add builder + Test + fix test
